backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.11.7
sse-starlette==2.1.3
httpx==0.28.1
pytest==9.0.3
langchain==0.3.27
llama-index==0.14.20
llama-index-vector-stores-milvus==1.1.0
pymilvus==2.6.12
langchain-openai==0.3.35
langchain-groq==0.3.8
python-dotenv==1.1.1
